requirements.txt
Mako==1.0.2
PyYAML==3.11
